Ranged Elite Enemy Implementation
Implemented the Ranged Elite Enemy
Has some weird bugs regarding the boundary of the play area. Also has the same model as the monochrome militia for now, but changing model should be easy. I also created a new testing scene called RangedEETests for all of my testing, and I also created a new prefab for the Monochrome militia, since the version we are using in the real combat scene is not a prefab and is instead just regular objects