Fixed Damage Dealt to Player
Edited PlayerHealth to make it so that it takes damage properly. Edited both MM and EE projectile scripts to make them deal the proper amount of damage (5 and 10 respectively. Then edited the DamageNumber UI Prefab to make it more dynamic and display the proper numbers (edited it's script as well). Finally edited the player prefab to make sure it's using the most up to date DamageNumber so that it works in the main scene